# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Local Test API",
        "description": "Simple CRUD RESTful API in Node.js",
        "version": "0.0.1",
        "license": {
            "name": "Apache License Version 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "http://172.17.100.20:3000",
            "description": "Local server"
        }
    ],
...
    "paths": {
...
        "/api/user/listusers?query={userEmail}&page={page}&pageSize={pageSize}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "getUserUsingGET",
                "summary": "Get user / users",
                "consumes": [
                    "application/json"
                ],
                "description": "Fetch the User object based on the provided query parameter. \"test@test.com\" for fetch one user or \"all\" for fetch all users.",
                "parameters": [
                    {
                        "name": "userEmail",
                        "required": true,
                        "in": "path",
                        "type": "string",
                        "description": "Email of the user"
                    },{
                        "name": "page",
                        "required": true,
                        "in": "path",
                        "type": "integer",
                        "description": "Pagina solicitada"
                    },{
                        "name": "pageSize",
                        "required": true,
                        "in": "path",
                        "type": "integer",
                        "description": "Tamanio de pagina solicitada"
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "$ref": "#/definitions/User"
                        },
                        "description": "A JSON array of user information"
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer", "example": 400
                                },
                                "message": {
                                    "type": "string", "example": "Invalid request body"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "409": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer", "example": 409
                                },
                                "message": {
                                    "type": "string", "example": "Requested item wasn't found!"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/user/users/{userEmail}": {
            "put": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "updateUserUsingPUT",
                "summary": "Update User",
                "consumes": [
                    "application/json"
                ],
                "description": "Update User object based on the provided object.",
                "parameters": [
                    {
                        "name": "userEmail",
                        "required": true,
                        "in": "path",
                        "type": "string",
                        "description": "Input the Email of the user which should be updated"
                    },
                    {
                        "in": "body",
                        "name": "updateUser",
                        "description": "Request body as following",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "userName": {
                                    "type": "string",
                                    "example": "test"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "test@test.com"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "role": {
                                    "type": "string",
                                    "example": "user"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer", "example": 400
                                },
                                "message": {
                                    "type": "string", "example": "Invalid request body"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "403": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer", "example": 403
                                },
                                "message": {
                                    "type": "string", "example": "Email is duplicate. You don't have permission to perform this operation!"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "409": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer", "example": 409
                                },
                                "message": {
                                    "type": "string", "example": "Requested email wasn't found!"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }, 
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "deleteUserUsingPUT",
                "summary": "Delete User",
                "consumes": [
                    "application/json"
                ],
                "description": "Delete User.",
                "parameters": [
                    {
                        "name": "userEmail",
                        "required": true,
                        "in": "path",
                        "type": "string",
                        "description": "Input the Email of the user which should be deleted"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "409": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer", "example": 409
                                },
                                "message": {
                                    "type": "string", "example": "Requested email wasn't found!"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/cupones/detalleCuponCompleto": {
            "post": {
                "tags": [
                    "cupon-controller"
                ],
                "operationId": "detalleCuponCompletoUsingGET",
                "summary": "Detalle del cupon",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "image/jpeg"
                ],
                "description": "Detalle de cupon y retorna la imagen relacionada desde S3   ",
                "parameters": [
                    {
                        "in": "body",
                        "name": "detalle",
                        "description": "Request body as following",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    },{
                        "in": "header",
                        "name": "authorization",
                        "description": "Access Token JWT",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVC..."
                        }
                    },
                    {
                        "in": "header",
                        "name": "refresh-token",
                        "description": "Refresh Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVC..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with cupon details and image",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "detalles": {
                                    "$ref": "#/definitions/Cupon"
                                },
                                "image": {
                                    "type": "string",
                                    "description": "Base64 encoded image content, prepend with 'data:image/jpeg;base64,' to display as image in HTML.",
                                    "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD..."
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer", "example": 400
                                },
                                "message": {
                                    "type": "string", "example": "Invalid request body"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "409": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer", "example": 409
                                },
                                "message": {
                                    "type": "string", "example": "Requested item wasn't found!"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "userName": {
                    "type": "string",
                    "example": "Jimmy"
                },
                "email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "createdAT": {
                    "type":"string",
                    "example": "2022-09-19T09:03:45.281Z"
                },
                "updatedAt": {
                    "type":"string",
                    "example": "2022-09-19T09:03:46.865Z"
                }
            }
        },
...
        "Cupon": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "fidLocatario": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "fidTipoCupon": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "sumilla": {
                    "type": "string",
                    "example": "20% DE DESCUENTO"
                },
                "descripcionCompleta": {
                    "type": "string",
                    "example": "Ten 20% de descuento en la compra de un corndog en Chung chun"
                },
                "fechaExpiracion": {
                    "type":"string",
                    "example": "2024-06-30 00:00:00"
                },
                "terminosCondiciones": {
                    "type": "string",
                    "example": "Válido solo en Chung chun de Plaza San Miguel, no acumulable con otras promociones"
                },
                "esLimitado": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "costoPuntos": {
                    "type": "integer",
                    "format": "int64",
                    "example": 120
                },
                "cantidadInicial": {
                    "type": "integer",
                    "format": "int64",
                    "example": 100
                },
                "cantidadDisponible": {
                    "type": "integer",
                    "format": "int64",
                    "example": 80
                },
                "ordenPriorizacion": {
                    "type": "integer",
                    "format": "int64",
                    "example": 3
                },
                "rutaFoto": {
                    "type": "string",
                    "example": "www.vsda"
                },
                "activo": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "createdAT": {
                    "type":"string",
                    "example": "2022-09-19T09:03:45.281Z"
                },
                "updatedAt": {
                    "type":"string",
                    "example": "2022-09-19T09:03:46.865Z"
                }
            }
        },
...
    }
}